Minor file structure changes
