+++ index.html
@@ -9,6 +9,7 @@
           <title>Document</title>
           <link rel="stylesheet" href="./dist/css/index.css" />
           <script src="./node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
+          <script type="module" src="./src/js/index.mjs"></script>
      </head>
      <body class="background-primary">
           <main class="container vh-100">
@@ -216,5 +217,4 @@
                </div>
           </main>
      </body>
-     <script type="module" src="./src/js/index.mjs"></script>
 </html>
