+++ src/js/index.mjs
@@ -1,5 +1,5 @@
 import { validateBootstrapForms } from "./utilities/forms.mjs";
-import { collapsableForm } from "./utilities/forms.mjs";
+import { collapsableForm } from "./utilities/formcollapse.mjs";
 
 validateBootstrapForms();
 collapsableForm();
